[micro] Generalize and reorg printing code
