+++ dpl_micropolicies/TestHeapSafety.hs
@@ -218,170 +218,3 @@ maxInstrsToGenerate = 10
 prop_noninterference :: PolicyPlus -> MStatePair -> Property
 prop_noninterference pplus ms = prop_NI' pplus 0 maxInstrsToGenerate [] ms
 
--------------------------------------------------------------------------------------
--- Printing for Heap Safety policy (but quite a bit should be generic!)
-
-prettyMStatePair :: PolicyPlus -> MStatePair -> Doc
-prettyMStatePair pplus (M (m1, p1) (m2, p2)) =
-    let ppol = policy pplus in
-    P.vcat [ P.text "PC:" <+> pretty pplus (f_pc m1, p_pc p1) (f_pc m2, p_pc p2)
-           , P.text "Registers:" $$ P.nest 2 (pretty pplus (f_gprs m1, p_gprs p1) (f_gprs m2, p_gprs p2))
-           , P.text "Memories:" $$ P.nest 2 (pretty pplus (f_mem m1, p_mem p1) (f_mem m2, p_mem p2))
-           , (compareMachines pplus) pplus $ M (m1,p1) (m2,p2)
-           ]
-
-print_mstatepair :: PolicyPlus -> MStatePair -> IO ()
-print_mstatepair pplus m = putStrLn $ P.render $ prettyMStatePair pplus m
-
-verboseTracing = False
---verboseTracing = True
-
-printTrace pplus tr1 tr2 = putStrLn $ P.render $ prettyTrace pplus tr1 tr2
-
-prettyTrace :: PolicyPlus -> [(Machine_State, PIPE_State)] -> [(Machine_State, PIPE_State)] -> Doc
-prettyTrace pplus [] [] = P.empty
-prettyTrace pplus [(m1,p1)] [(m2,p2)] = prettyMStatePair pplus (M (m1,p1) (m2,p2))
-prettyTrace pplus (tr1@((m1,p1):_)) (tr2@((m2,p2):_)) =
-    prettyMStatePair pplus (M (m1,p1) (m2,p2)) $$ P.text ""
-      $$ P.text "Trace:" $$ prettyDiffs pplus tr1 tr2
-
-prettyDiffs :: PolicyPlus -> [(Machine_State, PIPE_State)] -> [(Machine_State, PIPE_State)] -> Doc
-prettyDiffs pplus ((m11,p11):(m12,p12):tr1) ((m21,p21):(m22,p22):tr2) =
-  (if verboseTracing then
-       P.text "----------------------------------------------------------------"
-    $$ P.nest 10 (P.text "Raw Machine 1 memory:" $$ P.nest 3 (P.text (show $ f_dm $ f_mem m12)))
-    $$ P.nest 10 (P.text "Raw Machine 1 tags:" $$ P.nest 3 (P.text (show $ p_mem p12)))
-    $$ P.nest 10 (P.text "Raw Machine 2 memory:" $$ P.nest 3 (P.text (show $ f_dm $ f_mem m22)))
-    $$ P.nest 10 (P.text "Raw Machine 2 tags:" $$ P.nest 3 (P.text (show $ p_mem p22)))
-    $$ P.nest 10 (P.text "Machine 1:" $$ P.nest 3 (pretty pplus m12 p12) $$
-                  P.text "Machine 2" $$ P.nest 3 (pretty pplus m22 p22) )
-  else
-    P.empty)
-  $$ pretty pplus (calcDiff pplus (m11,p11) (m12,p12))
-                  (calcDiff pplus (m21,p21) (m22,p22))
-  $$ prettyDiffs pplus ((m12,p12):tr1) ((m22,p22):tr2)
-prettyDiffs pplus [(m1,p1)] [(m2,p2)] =
-  P.text "" $$ P.text "Final:" $$ prettyMStatePair pplus (M (m1,p1) (m2,p2))
-prettyDiffs _ _ _ = P.empty
-
-data Diff = Diff { d_pc :: (Integer, TagSet)               -- value and tag of the current PC
-                 , d_instr :: Maybe Instr_I                -- current instruction
-                 , d_reg :: [(GPR_Addr, Integer, TagSet)]  -- change in registers
-                 , d_mem :: [(Integer, Integer, TagSet)]   -- Change in memory
-                 }
-
--- Generic "find diffs" function: Takes two association lists l1 and
--- l2, both assumed sorted by their keys and both representing
--- *infinite* maps with some default value d (passed as third
--- parameter), and returns a list of changes
---
--- N.b. In the cases where we are returning something, we first have
--- to check whether the thing we are returning is equal to d!  (And
--- not return it in this case.)
-diff :: (Ord a, Eq b) => [(a, b)] -> [(a, b)] -> b -> [(a, (b, b))]
-diff [] [] d = []
-diff ((x1,y1):l1) [] d = (if y1==d then [] else [(x1,(y1,d))]) ++ diff l1 [] d
-diff [] ((x2,y2):l2) d = (if y2==d then [] else [(x2,(d,y2))]) ++ diff [] l2 d
-diff ((x1,y1):l1) ((x2,y2):l2) d
-         | x1 < x2   = (if y1==d then [] else [(x1,(y1,d))]) ++ diff l1 ((x2,y2):l2) d
-         | x1 > x2   = (if y2==d then [] else [(x2,(d,y2))]) ++ diff ((x1,y1):l1) l2 d
-         | otherwise = (if y1==y2 then [] else [(x1,(y1,y2))]) ++ diff l1 l2 d 
-
-calcDiff :: PolicyPlus -> (Machine_State, PIPE_State) -> (Machine_State, PIPE_State) -> Diff
-calcDiff pplus (m1,p1) (m2,p2) =
-  Diff {
-    d_pc = (f_pc m1, p_pc p1)
-  , d_instr =
-      case fst $ instr_fetch m1 of
-        Fetch u32 -> decode_I RV32 u32
-        _ -> error "Bad instr fetch in calcDiff"
-  , d_reg =
-      let GPR_File r1 = f_gprs m1
-          GPR_File r2 = f_gprs m2
-          GPR_FileT t1 = p_gprs p1
-          GPR_FileT t2 = p_gprs p2
-          reg_diff =
-            filter (\((i1,d1),(i2,d2)) -> assert (i1 == i2) $ d1 /= d2)
-                   (zip (Data_Map.assocs r1) (Data_Map.assocs r2))
-          tag_diff =
-            filter (\((i1,l1),(i2,l2)) -> assert (i1 == i2) $ l1 /= l2)
-                   (zip (Data_Map.assocs t1) (Data_Map.assocs t2))
-      in case (reg_diff, tag_diff) of
-           ([], []) -> []
-           ([((i,_),(_,d))],[((j,_),(_,l))]) | i == j -> [(i,d,l)]
-           ([((i,_),(_,d))],[]) ->
-             catMaybes [(i,d,) <$> Data_Map.lookup i t2]
-           ([],[((i,_),(_,l))]) ->
-             catMaybes [(i,,l) <$> Data_Map.lookup i r2]
-           _ -> -- TODO (Leo!)
-                error $ "More than one diff in register file:" ++
-                        " registers = " ++ show reg_diff ++
-                        " and tags = " ++ show tag_diff
-  , d_mem =
-      let Mem dm1 _ = f_mem m1
-          Mem dm2 _ = f_mem m2
-          MemT pm1 = p_mem p1
-          MemT pm2 = p_mem p2
-          both1 = map (\((i,d),(j,t)) -> assert (i==j) $ (i,(d,t))) $ zip (Data_Map.assocs dm1) (Data_Map.assocs pm1)
-          both2 = map (\((i,d),(j,t)) -> assert (i==j) $ (i,(d,t))) $ zip (Data_Map.assocs dm2) (Data_Map.assocs pm2)
-          diffs = diff both1 both2 (uninitialized_word, emptyInstTag pplus)
-          extract (i,(_,(d,t))) = (i,d,t)
-       in map extract diffs 
-  }
-
---          data_diff =
---            filter (\((i1,d1),(i2,d2)) ->
---                      if i1 == i2 then d1 /= d2 else error $ "DIFF: " ++ show ("i1", i1, "d1", d1, "i2", i2, "d2", d2, "dm1", dm1, "dm2", dm2))
-----                             assert (i1 == i2) $ d1 /= d2)
---                   (zip (Data_Map.assocs dm1) (Data_Map.assocs dm2))
---          tag_diff =
---            filter (\((i1,l1),(i2,l2)) -> assert (i1 == i2) $ l1 /= l2) (zip (Data_Map.assocs pm1) (Data_Map.assocs pm2))
---      in case (data_diff, tag_diff) of
---           ([], []) -> Nothing
---           ([((i,_),(_,d))],[((j,_),(_,l))]) | i == j -> Just (i,d,l)
---           ([((i,_),(_,d))],[]) ->
---             (i,d,) <$> Data_Map.lookup i pm2
---           ([],[((i,_),(_,l))]) ->
---             (i,,l) <$> Data_Map.lookup i dm2
---           _ -> error $ "More than one diff in memory file:" ++
---                        " data = " ++ show data_diff ++
---                        " and tags = " ++ show tag_diff
-
-prettyRegDiff pplus ((i,d,l):r1) ((i', d', l'):r2)
-    | i == i', d == d', l == l' =
-        (P.char 'r' P.<> P.integer i <+> P.text "<-" <+> pretty pplus d l)
-        $$ prettyRegDiff pplus r1 r2
-    | otherwise =
-      (ppStrong (P.char 'r' P.<> P.integer i <+> P.text "<-" <+> pretty pplus d l <||>
-                 P.char 'r' P.<> P.integer i' <+> P.text "<-" <+> pretty pplus d' l'))
-      $$ prettyRegDiff pplus r1 r2
-prettyRegDiff _ [] [] = P.empty
--- TODO (Leo): This can happen a lot now...
-prettyRegDiff _ r1 r2 = P.text $ "<prettyRegDiff??> " ++ show (r1,r2)
-
-prettyMemDiff pplus ((i,d,l):m1) ((i', d', l'):m2)
-    | i == i', d == d', l == l' =
-        (P.char '[' P.<> P.integer i P.<> P.char ']' <+> P.text "<-" <+> pretty pplus d l)
-        $$ prettyMemDiff pplus m1 m2
-    | otherwise =
-      (ppStrong (P.char '[' P.<> P.integer i P.<> P.char ']' <+> P.text "<-" <+> pretty pplus d l
-                 <||> P.char '[' P.<> P.integer i' P.<> P.char ']' <+> P.text "<-" <+> pretty pplus d' l'))
-      $$ prettyMemDiff pplus m1 m2
-prettyMemDiff _ [] [] = P.empty
-prettyMemDiff _ _ _ = P.text "<prettyMemDiff??>"
-
-instance CoupledPP (Maybe Instr_I) (Maybe Instr_I) where
-  pretty pplus (Just i1) (Just i2)
-    | i1 == i2  = pp pplus i1
-    | otherwise = ppStrong (pp pplus i1 <||> pp pplus i2)
-  pretty _ Nothing Nothing = P.text "<Bad instr>"
-
-instance CoupledPP Diff Diff where
-  pretty pplus d1 d2 =
-    P.hcat [ pad 17 (pretty pplus (d_pc d1) (d_pc d2))
-           , P.text " "
-           , pad 17 (pretty pplus (d_instr d1) (d_instr d2))
-           , P.text "     "
-           , prettyRegDiff pplus (d_reg d1) (d_reg d2)
-           , prettyMemDiff pplus (d_mem d1) (d_mem d2)
-           ]
